Version ready for testing on AUX01
The script now gets all the actions fromt he previous night
and has a record of the das machine+camera combinations. Unmounted
cameras have a das machine of None and are ignored. This must be updated
later as the cameras are mounted.

What we get at the end now is a list of images from each action.
When Simon finishes the new interface into create_images.py we can
run that with the list of images and the directory for the pngs.

This script's name will likely change later when I am done as get_actions
will not be the only thing it is doing.

JMCC